requirements.txt
streamlit
pandas
matplotlib
seaborn
scikit-learn
joblib
rdkit
shap
biopython
openai
streamlit-ketcher